refactor: 위시 관련 구조 수정
#66
