+++ src/main/java/com/tikklesaver/domain/wish/controller/WishController.java
@@ -3,12 +3,12 @@ package com.tikklesaver.domain.wish.controller;
 import com.tikklesaver.domain.product.entity.Product;
 import com.tikklesaver.domain.product.service.ProductCommandService;
 import com.tikklesaver.domain.wish.converter.WishConverter;
-import com.tikklesaver.domain.wish.dto.WishRequestDTO;
-import com.tikklesaver.domain.wish.dto.WishResponseDTO;
+import com.tikklesaver.domain.wish.dto.wish.WishRequestDTO;
+import com.tikklesaver.domain.wish.dto.wish.WishResponseDTO;
 import com.tikklesaver.domain.wish.entity.Wish;
 import com.tikklesaver.domain.wish.entity.enums.SatisfactionStatus;
-import com.tikklesaver.domain.wish.service.WishQueryService;
-import com.tikklesaver.domain.wish.service.WishCommandService;
+import com.tikklesaver.domain.wish.service.wish.WishQueryService;
+import com.tikklesaver.domain.wish.service.wish.WishCommandService;
 import com.tikklesaver.global.apiPayload.ApiResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
