+++ src/main/java/com/tikklesaver/domain/wish/repository/wish/WishRepository.java
@@ -1,4 +1,4 @@
-package com.tikklesaver.domain.wish.repository;
+package com.tikklesaver.domain.wish.repository.wish;
 
 import com.tikklesaver.domain.wish.entity.Wish;
 import org.springframework.data.jpa.repository.JpaRepository;
