+++ src/main/java/com/tikklesaver/domain/wish/service/wish/WishQueryService.java
@@ -1,6 +1,6 @@
-package com.tikklesaver.domain.wish.service;
+package com.tikklesaver.domain.wish.service.wish;
 
-import com.tikklesaver.domain.wish.dto.WishResponseDTO;
+import com.tikklesaver.domain.wish.dto.wish.WishResponseDTO;
 
 import java.util.List;
